COZMO-7354: Enable implicit conversion warnings (#2946)
* COZMO-7354: Enable implicit conversion warnings

Implicit conversion warnings will appear but do not fail the build.
Exposes

Developers will be responsible for fixing warnings in their area.

If we like these settings, they can be applied to other projects.

* COZMO-7354: Enable -Wconversion for android

Xcode setting (CLANG_WARN_SUSPICIOUS_IMPLICIT_CONVERSION) was not
reflected by android build command.  Add -Wconversion to compiler flags
to enable on all platforms.

* COZMO-7354: Enable implicit conversion warnings

Implicit conversion warnings will appear but do not fail the build.
Exposes

Developers will be responsible for fixing warnings in their area.

If we like these settings, they can be applied to other projects.

* COZMO-7354: Enable -Wconversion for android

Xcode setting (CLANG_WARN_SUSPICIOUS_IMPLICIT_CONVERSION) was not
reflected by android build command.  Add -Wconversion to compiler flags
to enable on all platforms.

* COZMO-7354: Enable implicit conversion warnings

Implicit conversion warnings will appear but do not fail the build.
Exposes

Developers will be responsible for fixing warnings in their area.

If we like these settings, they can be applied to other projects.

* COZMO-7354: Enable implicit conversion warnings

Implicit conversion warnings will appear but do not fail the build.
Exposes

Developers will be responsible for fixing warnings in their area.

If we like these settings, they can be applied to other projects.

* COZMO-7354: Enable -Wconversion for android

Xcode setting (CLANG_WARN_SUSPICIOUS_IMPLICIT_CONVERSION) was not
reflected by android build command.  Add -Wconversion to compiler flags
to enable on all platforms.

* COZMO-7354: Enable -Wconversion for android

Xcode setting (CLANG_WARN_SUSPICIOUS_IMPLICIT_CONVERSION) was not
reflected by android build command.  Add -Wconversion to compiler flags
to enable on all platforms.

* COZMO-7354: Suppress implicit sign conversion warnings

Suppress sign conversion warnings to eliminate numerous false positives.
If this change is approved, I will open a new ticket to review sign
conversions on a future sprint.

* COZMO-7354: Fix float-to-bool conversion warnings

* COZMO-7354: Prefer numeric_cast over generic cast

Fix as requested by review.

* COZMO-7354: Various conversion fixes

Clean up high-volume conversion warnings in header files.

* COZMO-7354: Revise console variable ToggleValue()

I changed ToggleValue() to use cast operator provided by the type, if
any. I think this is the best fit for the intent of this method.

* COZMO-7354: Fix numerous implicit conversion warnings

I fixed numerous warnings that seemed like obvious wins.

I used Util::numeric_cast<> where it seemed appropriate, but I used
default (static) casts for low-level bit twiddles that don’t need to be
checked. I don’t want to introduce performance overhead in low-level
code.

Mac debug build reports 826 warnings left.

* COZMO-7354: Revert controversial warning flags

+++ basestation/include/anki/cozmo/basestation/robot.h
@@ -645,11 +645,25 @@ public:
   void ShiftEyes(AnimationStreamer::Tag& tag, f32 xPix, f32 yPix,
                  TimeStamp_t duration_ms, const std::string& name = "ShiftEyes");
   
-  AnimationStreamer& GetAnimationStreamer() { return _animationStreamer; }
-  void SetNumAnimationBytesPlayed(s32 numAnimationsBytesPlayed) { _numAnimationBytesPlayed = numAnimationsBytesPlayed; }
-  void SetNumAnimationAudioFramesPlayed(s32 numAnimationAudioFramesPlayed) { _numAnimationAudioFramesPlayed = numAnimationAudioFramesPlayed; }
-  void SetEnabledAnimTracks(s32 enabledAnimTracks) { _enabledAnimTracks = enabledAnimTracks; }
-  void SetAnimationTag(s32 animationTag) { _animationTag = animationTag; }
+  AnimationStreamer& GetAnimationStreamer() {
+    return _animationStreamer;
+  }
+  
+  void SetNumAnimationBytesPlayed(s32 numAnimationsBytesPlayed) {
+    _numAnimationBytesPlayed = numAnimationsBytesPlayed;
+  }
+  
+  void SetNumAnimationAudioFramesPlayed(s32 numAnimationAudioFramesPlayed) {
+    _numAnimationAudioFramesPlayed = numAnimationAudioFramesPlayed;
+  }
+  
+  void SetEnabledAnimTracks(u8 enabledAnimTracks) {
+    _enabledAnimTracks = enabledAnimTracks;
+  }
+  
+  void SetAnimationTag(u8 animationTag) {
+    _animationTag = animationTag;
+  }
 
   // Returns true if the robot is currently playing an animation, according
   // to most recent state message. NOTE: Will also be true if the animation
