+++ coretech/vision/include/anki/vision/basestation/colorPixelTypes.h
@@ -79,7 +79,7 @@ namespace Vision {
     
     // Convert to gray
     u8 gray() const {
-      u16 gray = r() + (g() << 1) + b(); // give green double weight
+      u16 gray = static_cast<u16>(r() + (g() << 1) + b()); // give green double weight
       gray = gray >> 2; // divide by 4
       assert(gray <= std::numeric_limits<u8>::max());
       return static_cast<u8>(gray);
@@ -145,7 +145,7 @@ namespace Vision {
   
   template<>
   inline u8 PixelRGB_<u8>::gray() const {
-    u16 gray = r() + (g() << 1) + b(); // give green double weight
+    u16 gray = static_cast<u16>(r() + (g() << 1) + b()); // give green double weight
     gray = gray >> 2; // divide by 4
     assert(gray <= std::numeric_limits<u8>::max());
     return static_cast<u8>(gray);
